% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{get_ggNetView_data}
\alias{get_ggNetView_data}
\title{Safely fetch a built-in dataset from ggNetView by name}
\usage{
get_ggNetView_data(name)
}
\description{
Safely fetch a built-in dataset from ggNetView by name
}
\keyword{internal}
