% Generated manually
\name{ggNetView.shiny-package}
\alias{ggNetView.shiny}
\alias{ggNetView.shiny-package}
\docType{package}
\title{Interactive Shiny Interface for 'ggNetView'}
\description{
A \pkg{shinydashboard}-based GUI for the \pkg{ggNetView} R package.
Use \code{\link{launch_ggNetView}} to start the application or
\code{\link{ggNetViewApp}} to obtain a \code{shinyApp} object for
deployment.
}
\seealso{
\code{\link[ggNetView]{ggNetView}}
}
\keyword{package}
