% Generated manually
\name{ggNetViewApp}
\alias{ggNetViewApp}
\title{Build the ggNetView Shiny app object}
\usage{
ggNetViewApp()
}
\value{
A \code{\link[shiny]{shinyApp}} object.
}
\description{
Returns a \code{shiny.appobj} that can be deployed on shinyapps.io,
Shiny Server or Posit Connect, instead of running the app directly.
}
