% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{is_graph_obj}
\alias{is_graph_obj}
\title{Detect whether x is a `tbl_graph` / `igraph` object}
\usage{
is_graph_obj(x)
}
\description{
Detect whether x is a `tbl_graph` / `igraph` object
}
\keyword{internal}
