% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_io.R
\name{list_ggNetView_data}
\alias{list_ggNetView_data}
\title{Safely list available built-in datasets in ggNetView}
\usage{
list_ggNetView_data()
}
\description{
Safely list available built-in datasets in ggNetView
}
\keyword{internal}
